Remove deprecated AddCommandParser
