+++ src/main/java/seedu/address/logic/parser/add/AddCommandParser.java
@@ -1,56 +1,44 @@
 package seedu.address.logic.parser.add;
 
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.clisyntax.ItemCliSyntax.PREFIX_ADDRESS;
-import static seedu.address.logic.parser.clisyntax.ItemCliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.parser.clisyntax.ItemCliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.clisyntax.ItemCliSyntax.PREFIX_PHONE;
-import static seedu.address.logic.parser.clisyntax.ItemCliSyntax.PREFIX_TAG;
-import static seedu.address.logic.parser.util.GeneralParserUtil.argumentsAreValid;
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_ITEM_TYPE;
+import static seedu.address.logic.parser.util.GeneralParserUtil.getCommandDetails;
+import static seedu.address.logic.parser.util.GeneralParserUtil.getItemType;
+import static seedu.address.model.util.ItemUtil.APPLICATION_ALIAS;
+import static seedu.address.model.util.ItemUtil.COMPANY_ALIAS;
+import static seedu.address.model.util.ItemUtil.INTERNSHIP_ALIAS;
+import static seedu.address.model.util.ItemUtil.PROFILE_ALIAS;
 
-import java.util.Set;
-
-import seedu.address.logic.commands.add.AddCommand;
-import seedu.address.logic.parser.ArgumentMultimap;
-import seedu.address.logic.parser.ArgumentTokenizer;
+import seedu.address.logic.commands.add.AddCommandAbstract;
 import seedu.address.logic.parser.Parser;
-import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.Phone;
-import seedu.address.model.tag.Tag;
 
 /**
- * Parses input arguments and creates a new AddCommand object.
+ * Parses input arguments for item type and returns AddCommandAbstract Object.
  */
-public class AddCommandParser implements Parser<AddCommand> {
+public class AddCommandParser implements Parser<AddCommandAbstract> {
 
     /**
-     * Parses the given {@code String} of arguments in the context of the AddCommand
+     * Parses the given {@code String} of arguments for the item type
      * and returns an AddCommand object for execution.
      *
-     * @throws ParseException if the user input does not conform the expected format.
+     * @throws ParseException if the user input does not conformClear the expected format
      */
-    public AddCommand parse(String args) throws ParseException {
-        ArgumentMultimap argMultimap =
-                ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL, PREFIX_ADDRESS, PREFIX_TAG);
+    public AddCommandAbstract parse(String args) throws ParseException {
+        String itemType = getItemType(args, AddCommandAbstract.MESSAGE_USAGE);
+        String commandDetails = getCommandDetails(args);
 
-        if (!argumentsAreValid(false, argMultimap, PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL, PREFIX_ADDRESS)) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, AddCommand.MESSAGE_USAGE));
+        switch (itemType) {
+        case COMPANY_ALIAS:
+            return new AddCompanyCommandParser().parse(commandDetails);
+        case INTERNSHIP_ALIAS:
+            return new AddInternshipCommandParser().parse(commandDetails);
+        case APPLICATION_ALIAS:
+            return new AddApplicationCommandParser().parse(commandDetails);
+        case PROFILE_ALIAS:
+            return new AddProfileCommandParser().parse(commandDetails);
+        default:
+            throw new ParseException(MESSAGE_INVALID_ITEM_TYPE);
         }
-
-        Name name = ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get());
-        Phone phone = ParserUtil.parsePhone(argMultimap.getValue(PREFIX_PHONE).get());
-        Email email = ParserUtil.parseEmail(argMultimap.getValue(PREFIX_EMAIL).get());
-        Address address = ParserUtil.parseAddress(argMultimap.getValue(PREFIX_ADDRESS).get());
-        Set<Tag> tagList = ParserUtil.parseTags(argMultimap.getAllValues(PREFIX_TAG));
-
-        Person person = new Person(name, phone, email, address, tagList);
-
-        return new AddCommand(person);
     }
 
 }
